refactor: comprehensive codebase cleanup and restructuring
- Replaced OrderedDict with dict (Python 3.7+ maintains insertion order)
- Standardized path handling using Path.as_posix() throughout
- Moved pipeline.py to xbbg/utils/ package for better organization
- Replaced namedtuple with dataclass for better type safety (ToQuery, CurrencyPair, Session)
- Updated all __init__.py files with proper docstrings and exports
- Improved exception handling specificity where appropriate
- Added type hints to _to_gen_ function
- Replaced os.path.exists with Path.exists()
- Fixed None comparisons (== None -> is None)
- Cleaned up code formatting and spacing
- Removed unnecessary comments while preserving useful ones

All changes maintain backward compatibility and pass linting checks.
