+++ xbbg/api/utils.py
@@ -1,0 +1,73 @@
+"""Bloomberg API utility functions.
+
+Provides utility functions for data processing and transformation.
+"""
+
+from __future__ import annotations
+
+import logging
+
+import pandas as pd
+
+logger = logging.getLogger(__name__)
+
+__all__ = ['adjust_ccy']
+
+
+def adjust_ccy(data: pd.DataFrame, ccy: str = 'USD') -> pd.DataFrame:
+    """Adjust series to a target currency.
+
+    Args:
+        data: daily price / turnover / etc. to adjust
+        ccy: currency to adjust to
+
+    Returns:
+        pd.DataFrame
+    """
+    from xbbg.api.historical import bdh  # noqa: PLC0415
+    from xbbg.api.reference import bdp  # noqa: PLC0415
+
+    if data.empty: return pd.DataFrame()
+    if ccy.lower() == 'local': return data
+    tickers = list(data.columns.get_level_values(level=0).unique())
+    start_date = data.index[0]
+    end_date = data.index[-1]
+
+    uccy = bdp(tickers=tickers, flds='crncy')
+    if not uccy.empty:
+        adj = (
+            uccy.crncy
+            .map(lambda v: {
+                'ccy': None if v.upper() == ccy else f'{ccy}{v.upper()} Curncy',
+                'factor': 100. if v[-1].islower() else 1.,
+            })
+            .apply(pd.Series)
+            .dropna(subset=['ccy'])
+        )
+    else: adj = pd.DataFrame()
+
+    if not adj.empty:
+        fx = (
+            bdh(tickers=adj.ccy.unique(), start_date=start_date, end_date=end_date)
+            .xs('Last_Price', axis=1, level=1)
+        )
+    else: fx = pd.DataFrame()
+
+    return (
+        pd.concat([
+            pd.Series(
+                (
+                    data[t]
+                    .dropna()
+                    .prod(axis=1)
+                    .div(
+                        (fx[adj.loc[t, 'ccy']] * adj.loc[t, 'factor'])
+                        if t in adj.index else 1.,
+                    )
+                ),
+                name=t,
+            )
+            for t in tickers
+        ], axis=1)
+    )
+
