+++ xbbg/io/__init__.py
@@ -1,0 +1,10 @@
+"""I/O utilities for file operations, caching, and storage.
+
+This package contains internal utilities for:
+- files: File system operations and path utilities
+- storage: Path resolution for cached Bloomberg data files
+- cached: Cache lookup and management for BDP/BDS queries
+- db: SQLite database wrapper for trial tracking
+- param: YAML configuration file loading and caching
+"""
+
